Modularising OverrideProcessor
